// src/modules/product/enums/product-message.enum.ts
export enum ProductMessage {
  CREATED = 'Product successfully created.',
  RETRIEVED_ALL = 'List of all products retrieved successfully.',
  RETRIEVED_ONE = 'Product retrieved successfully.',
  UPDATED = 'Product updated successfully.',
  DELETED = 'Product deleted successfully.',
  NOT_FOUND = 'Product not found.',
  BARCODE_NOT_FOUND = 'Product not found with this barcode.',
  CATEGORY_NOT_FOUND = 'Category not found.',
  SUBCATEGORY_NOT_FOUND = 'Subcategory not found.',
}